requirements.txt
aiocache==0.12.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
click==8.1.3
disnake==2.8.1
dnspython==2.3.0
fastapi==0.95.2
frozenlist==1.3.3
h11==0.14.0
idna==3.4
motor==3.1.2
multidict==6.0.4
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
yarl==1.9.2
